requirements.txt
streamlit==1.28.1
pandas==2.1.3
cryptography==41.0.7
torch==2.1.1
numpy
pathlib
requests
transformers==4.35.2
scikit-learn
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3
reportlab==4.0.7
plotly==5.17.0
